requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1

# ML Dependencies
scikit-learn==1.5.0
joblib==1.4.2
pandas==2.2.0
numpy==1.26.4